Chg: SettingsForm to ConfigureForm
